perf(parquet): Uses two writers instead of one
The Parquet encoding is a bottleneck because there are many data
generator threads but only a single writer. The generators produce
as fast as possible but is blocked by the writer encoding to Parquet
format.

This is a CPU bound workload and not disk I/O bound. For 10M the
created parquet file has a total size of under 400MB. The work stalls
when data generator threads are blocking on the single writer thread.

This is also a proof-of-concept to show the performance impact of
spreading write across two writers.

- Wall clock time reduced by 23%: 3.85s -> 2.89s
- CPU utliziation increased (less blocking)