refactor(parquet): disintermediate RecordBatch generation
Simplify control flow and reduce complexity of RecordBatch
generation significantly. It is not necessary anymore to flush
a RecordBatch when there are not enough rows by the caller. The
caller API is simple, just get a RecordBatch by specifying the
row count.

This refactoring removes `PartialContact` struct and allocating
an intermediate `Vec<PartialContact>` which is then converted to
a `RecordBatch`. The current approach fuses data generation with
building the RecordBatch into an efficient batch process.