Revert "feat(parquet): string interning generated names"
This reverts commit e895b92e7e43c5bf4f3964d211f94eb08d4b5ad4.

I am reverting this because it ended up degrading performance
significantly.
- IPC halved 2.08 to 1.10 instructions per cycle
- Cache misses doubled from 11% to 22%

perf stats summary:
33,883,545,682   instructions # 1.10  insn per cycle
146,031,921      cache-misse  # 22.49% of all cache refs

String interning is a good idea because of the many small heap
allocations which happens in `generate_name`. The problem is the
shared interner data structure between all threads.

In theory this may have worked. But in practice cache coherency
issues from all threads trying to simultaneously read & write to
the string interner data structure, blew up performance. The cost
of cross-core cache invalidations turned out to be much higher
than the savings from original allocation of 8million strings.

The next step is to try a thread-local to avoid cache coherence
problems. The impact is going to be smaller compared because the
interner is now local, rather than global. So across threads there
are going to be duplicate allocations. But this is worth a shot
because I expect it to scale linearly or better for workloads with
runs higher than 10M.