feat(tantivy): list directory files and contents
The `open_read` method is used for reading index data files as a
stream of chunks. This is good as the entire file does not have
to be materialized into a memory byte array all at once.

To read the metadata JSON file: `meta.json`, we have to use the
`atomic_read` method. Using `open_read` will result in an error
because it checks for the existence of magic bytes in the footer
which does not apply to a JSON file.

This sets up the next part which is to efficiently extract all
the directory contents into a byte array with spill to disk so
that the build phase is dependent on system memory.